requirements.txt
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
fastapi==0.61.1
h11==0.11.0
mysql-connector-python==8.0.22
passlib==1.7.4
protobuf==3.13.0
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.12.2
